Consolidate parallel fragment e2e tests into single file pair
Merged test scenarios from:
- st_fragment_parallel_api_restrictions.py
- st_fragment_parallel_api_restrictions_test.py
- st_fragment_parallel_cancellation.py
- st_fragment_parallel_cancellation_test.py

Into the base files:
- st_fragment_parallel.py (with test_mode query param routing)
- st_fragment_parallel_test.py (with all test functions)

This follows the project's e2e test guidelines to group related tests
into single, logical test files for CI efficiency.
