+++ e2e_playwright/st_fragment_parallel_test.py
@@ -20,8 +20,12 @@ import re
 
 from playwright.sync_api import Page, expect
 
-from e2e_playwright.conftest import wait_for_app_run
+from e2e_playwright.conftest import build_app_url, wait_for_app_run
 from e2e_playwright.shared.app_utils import click_button, get_element_by_key
+
+# =============================================================================
+# Core parallel fragment tests (use `app` fixture - auto-navigates to default mode)
+# =============================================================================
 
 
 def test_parallel_fragments_render_concurrently(app: Page) -> None:
@@ -115,3 +119,155 @@ def test_parallel_fragment_container_matches_main_thread(app: Page) -> None:
         block = vertical_blocks.nth(i)
         inner_text = block.inner_text()
         assert inner_text.strip() != "", f"Found empty container at index {i}"
+
+
+# =============================================================================
+# API Restrictions tests (use `page` + `app_base_url` fixtures for explicit navigation)
+# =============================================================================
+
+
+def test_parallel_fragment_blocks_dialog_during_initial_run(
+    page: Page, app_base_url: str
+) -> None:
+    """Parallel fragment unconditionally calling @st.dialog shows error."""
+    page.goto(build_app_url(app_base_url, query="test=dialog_block"))
+    wait_for_app_run(page)
+
+    exception_message = page.get_by_test_id("stExceptionMessage")
+    expect(exception_message).to_be_visible()
+    expect(exception_message).to_contain_text(
+        "cannot be called from a parallel fragment"
+    )
+    expect(exception_message).to_contain_text("@st.dialog")
+
+
+def test_parallel_fragment_blocks_switch_page_during_initial_run(
+    page: Page, app_base_url: str
+) -> None:
+    """Parallel fragment calling st.switch_page shows error."""
+    page.goto(build_app_url(app_base_url, query="test=switch_page_block"))
+    wait_for_app_run(page)
+
+    exception_message = page.get_by_test_id("stExceptionMessage")
+    expect(exception_message).to_be_visible()
+    expect(exception_message).to_contain_text(
+        "cannot be called from a parallel fragment"
+    )
+    expect(exception_message).to_contain_text("st.switch_page")
+
+
+def test_parallel_fragment_allows_dialog_on_rerun(
+    page: Page, app_base_url: str
+) -> None:
+    """Parallel fragment allows dialog when triggered by button click."""
+    page.goto(build_app_url(app_base_url, query="test=dialog_allow_rerun"))
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Fragment content")).to_be_visible()
+
+    click_button(page, "Open Dialog")
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Dialog opened successfully")).to_be_visible()
+
+
+def test_nested_sequential_fragment_blocks_dialog_during_parallel_batch(
+    page: Page, app_base_url: str
+) -> None:
+    """Non-parallel fragment nested inside parallel fragment inherits restriction."""
+    page.goto(build_app_url(app_base_url, query="test=nested_sequential_block"))
+    wait_for_app_run(page)
+
+    # There may be multiple exception messages (original + wrapped), use first()
+    exception_message = page.get_by_test_id("stExceptionMessage").first
+    expect(exception_message).to_be_visible()
+    expect(exception_message).to_contain_text(
+        "cannot be called from a parallel fragment"
+    )
+
+
+def test_nested_parallel_fragments_both_restricted(
+    page: Page, app_base_url: str
+) -> None:
+    """Parallel fragment nested inside parallel fragment is also restricted."""
+    page.goto(build_app_url(app_base_url, query="test=nested_parallel_block"))
+    wait_for_app_run(page)
+
+    exception_message = page.get_by_test_id("stExceptionMessage")
+    expect(exception_message).to_be_visible()
+    expect(exception_message).to_contain_text(
+        "cannot be called from a parallel fragment"
+    )
+
+
+def test_nested_parallel_fragment_allows_dialog_on_rerun(
+    page: Page, app_base_url: str
+) -> None:
+    """Nested parallel fragment allows dialog on rerun."""
+    page.goto(build_app_url(app_base_url, query="test=nested_parallel_allow_rerun"))
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Inner fragment")).to_be_visible()
+
+    click_button(page, "Open Nested Dialog")
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Nested dialog opened successfully")).to_be_visible()
+
+
+# =============================================================================
+# Cancellation tests (use `page` + `app_base_url` fixtures for explicit navigation)
+# =============================================================================
+
+
+def test_parallel_st_stop_ends_script(page: Page, app_base_url: str) -> None:
+    """Fragment A calls st.stop, should end quickly without waiting for Fragment B."""
+    page.goto(build_app_url(app_base_url, query="test=st_stop"))
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Fragment A content")).to_be_visible()
+    # Fragment B should not complete due to st.stop in Fragment A
+    expect(page.get_by_text("Fragment B done after 5s")).not_to_be_visible()
+
+    total_time_text = page.get_by_text(re.compile(r"Total time: \d+\.\d+s"))
+    expect(total_time_text).to_be_visible()
+
+
+def test_parallel_st_rerun_restarts_app(page: Page, app_base_url: str) -> None:
+    """Fragment calls st.rerun on first run, app restarts."""
+    page.goto(build_app_url(app_base_url, query="test=st_rerun"))
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("App restarted successfully")).to_be_visible()
+    # Run 1: run_count starts at 0, incremented to 1, fragment sets to 2 and reruns
+    # Run 2: run_count is 2, incremented to 3, fragment shows success
+    expect(page.get_by_text("Run count: 3", exact=True)).to_be_visible()
+
+
+def test_widget_interaction_during_parallel_execution(
+    page: Page, app_base_url: str
+) -> None:
+    """Button click in fast fragment while slow fragment is running."""
+    page.goto(build_app_url(app_base_url, query="test=widget_interaction"))
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Counter: 0", exact=True)).to_be_visible()
+
+    click_button(page, "Increment")
+    wait_for_app_run(page)
+
+    expect(page.get_by_text("Counter: 1", exact=True)).to_be_visible()
+
+
+def test_parallel_fragment_error_renders_in_container(
+    page: Page, app_base_url: str
+) -> None:
+    """Error in Fragment A renders in A's container, Fragment B succeeds."""
+    page.goto(build_app_url(app_base_url, query="test=error_container"))
+    wait_for_app_run(page)
+
+    exception_message = page.get_by_test_id("stExceptionMessage")
+    expect(exception_message).to_be_visible()
+    expect(exception_message).to_contain_text("ValueError")
+    expect(exception_message).to_contain_text("Test error in fragment")
+    expect(page.get_by_text("Fragment B success")).to_be_visible()
